Lido app/marcus/json (#55)
Changed Error format to a JSON
