+++ package.json
@@ -42,11 +42,13 @@
     "docdash": "^1.2.0",
     "format": "^0.2.2",
     "gh-pages": "^3.1.0",
+    "jscodeshift": "^0.13.0",
     "jsdoc": "^3.6.5",
     "jstat": "^1.9.3",
     "mocha": "^7.2.0",
     "nyc": "^15.1.0",
     "prettier": "^2.5.1",
+    "sinon": "^12.0.1",
     "webpack": "^4.44.0",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-cli": "^3.3.12",
